Rover-Django/requirements.txt
# db
psycopg2
psycopg2-binary

# testing
model-mommy

# local
django-extensions